requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
Django==2.1.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
importlib-metadata==1.4.0
kombu==4.6.7
more-itertools==8.1.0
mysqlclient==1.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
vine==1.3.0
zipp==0.6.0
